+++ tests/conftest.py
@@ -49,6 +49,7 @@ except ImportError:
         NUMBER = "number"
         SELECT = "select"
         BUTTON = "button"
+        UPDATE = "update"
 
     _const.Platform = _Platform  # type: ignore[attr-defined]
     _const.PERCENTAGE = "%"  # type: ignore[attr-defined]
@@ -348,6 +349,28 @@ except ImportError:
     _button.ButtonEntity = _ButtonEntity  # type: ignore[attr-defined]
     sys.modules.setdefault("homeassistant.components.button", _button)
 
+    _update = types.ModuleType("homeassistant.components.update")
+
+    class _UpdateDeviceClass(str, enum.Enum):
+        FIRMWARE = "firmware"
+
+    class _UpdateEntity:
+        _attr_release_summary = None
+
+        @property
+        def release_summary(self):
+            return self._attr_release_summary
+
+        def async_on_remove(self, func):
+            self._unsub = func
+
+        def async_write_ha_state(self):
+            return None
+
+    _update.UpdateDeviceClass = _UpdateDeviceClass  # type: ignore[attr-defined]
+    _update.UpdateEntity = _UpdateEntity  # type: ignore[attr-defined]
+    sys.modules.setdefault("homeassistant.components.update", _update)
+
     _cover = types.ModuleType("homeassistant.components.cover")
     _cover.ATTR_POSITION = "position"  # type: ignore[attr-defined]
 
@@ -425,6 +448,12 @@ except ImportError:
             super().__init__(**kwargs)
 
     _dr.DeviceInfo = _DeviceInfo  # type: ignore[attr-defined]
+    _dr.EVENT_DEVICE_REGISTRY_UPDATED = "device_registry_updated"  # type: ignore[attr-defined]
+
+    def _dr_async_get(hass):
+        return getattr(hass, "device_registry", None)
+
+    _dr.async_get = _dr_async_get  # type: ignore[attr-defined]
     sys.modules.setdefault("homeassistant.helpers.device_registry", _dr)
 
     _ep = types.ModuleType("homeassistant.helpers.entity_platform")